+++ radioactive/__main__.py
@@ -8,11 +8,11 @@ from zenlog import log
 
 from radioactive.alias import Alias
 from radioactive.app import App
+from radioactive.ffplay import Ffplay, kill_background_ffplays
 from radioactive.handler import Handler
 from radioactive.help import show_help
 from radioactive.last_station import Last_station
 from radioactive.parser import parse_options
-from radioactive.player import Player, kill_background_ffplays
 from radioactive.utilities import (
     check_sort_by_parameter,
     handle_add_station,
@@ -22,6 +22,7 @@ from radioactive.utilities import (
     handle_favorite_table,
     handle_listen_keypress,
     handle_play_last_station,
+    handle_play_random_station,
     handle_record,
     handle_save_last_station,
     handle_search_stations,
@@ -34,20 +35,43 @@ from radioactive.utilities import (
 
 # globally needed as signal handler needs it
 # to terminate main() properly
+ffplay = None
 player = None
 
 
 def final_step(options, last_station, alias, handler):
+    global ffplay  # always needed
     global player
+
     # check target URL for the last time
     if options["target_url"].strip() == "":
         log.error("something is wrong with the url")
         sys.exit(1)
 
+    if options["audio_player"] == "vlc":
+        from radioactive.vlc import VLC
+
+        vlc = VLC()
+        vlc.start(options["target_url"])
+        player = vlc
+
+    elif options["audio_player"] == "mpv":
+        from radioactive.mpv import MPV
+
+        mpv = MPV()
+        mpv.start(options["target_url"])
+        player = mpv
+
+    elif options["audio_player"] == "ffplay":
+        ffplay = Ffplay(options["target_url"], options["volume"], options["loglevel"])
+        player = ffplay
+
+    else:
+        log.error("Unsupported media player selected")
+        sys.exit(1)
+
     if options["curr_station_name"].strip() == "":
         options["curr_station_name"] = "N/A"
-
-    player = Player(options["target_url"], options["volume"], options["loglevel"])
 
     handle_save_last_station(
         last_station, options["curr_station_name"], options["target_url"]
@@ -72,6 +96,7 @@ def final_step(options, last_station, alias, handler):
 
     handle_listen_keypress(
         alias,
+        player,
         target_url=options["target_url"],
         station_name=options["curr_station_name"],
         station_url=options["target_url"],
@@ -89,8 +114,6 @@ def main():
 
     options = parse_options()
 
-    handle_welcome_screen()
-
     VERSION = app.get_version()
 
     handler = Handler()
@@ -103,6 +126,8 @@ def main():
     if options["version"]:
         log.info("RADIO-ACTIVE : version {}".format(VERSION))
         sys.exit(0)
+
+    handle_welcome_screen()
 
     if options["show_help_table"]:
         show_help()
@@ -134,7 +159,10 @@ def main():
     # ----------- country ----------- #
     if options["discover_country_code"]:
         response = handler.discover_by_country(
-            options["discover_country_code"], options["limit"], options["sort_by"]
+            options["discover_country_code"],
+            options["limit"],
+            options["sort_by"],
+            options["filter_with"],
         )
         if response is not None:
             (
@@ -148,7 +176,10 @@ def main():
     # -------------- state ------------- #
     if options["discover_state"]:
         response = handler.discover_by_state(
-            options["discover_state"], options["limit"], options["sort_by"]
+            options["discover_state"],
+            options["limit"],
+            options["sort_by"],
+            options["filter_with"],
         )
         if response is not None:
             (
@@ -162,7 +193,10 @@ def main():
     # ----------- language ------------ #
     if options["discover_language"]:
         response = handler.discover_by_language(
-            options["discover_language"], options["limit"], options["sort_by"]
+            options["discover_language"],
+            options["limit"],
+            options["sort_by"],
+            options["filter_with"],
         )
         if response is not None:
             (
@@ -176,7 +210,10 @@ def main():
     # -------------- tag ------------- #
     if options["discover_tag"]:
         response = handler.discover_by_tag(
-            options["discover_tag"], options["limit"], options["sort_by"]
+            options["discover_tag"],
+            options["limit"],
+            options["sort_by"],
+            options["filter_with"],
         )
         if response is not None:
             (
@@ -193,6 +230,7 @@ def main():
         and options["search_station_uuid"] is None
         and options["direct_play"] is None
         and not options["play_last_station"]
+        and not options["play_random"]
     ):
         (
             options["curr_station_name"],
@@ -221,6 +259,7 @@ def main():
             options["search_station_name"],
             options["limit"],
             options["sort_by"],
+            options["filter_with"],
         )
         if response is not None:
             (
@@ -237,6 +276,13 @@ def main():
         options["curr_station_name"], options["target_url"] = handle_direct_play(
             alias, options["direct_play"]
         )
+        final_step(options, last_station, alias, handler)
+
+    if options["play_random"]:
+        (
+            options["curr_station_name"],
+            options["target_url"],
+        ) = handle_play_random_station(alias)
         final_step(options, last_station, alias, handler)
 
     if options["play_last_station"]:
@@ -259,11 +305,15 @@ def main():
 
 
 def signal_handler(sig, frame):
+    global ffplay
     global player
     log.debug("You pressed Ctrl+C!")
     log.debug("Stopping the radio")
-    if player and player.is_playing:
+    if ffplay and ffplay.is_playing:
+        ffplay.stop()
+        #  kill the player
         player.stop()
+
     log.info("Exiting now")
     sys.exit(0)