+++ radioactive/parser.py
@@ -29,8 +29,10 @@ def parse_options():
 
     options["play_last_station"] = args.play_last_station
     options["direct_play"] = args.direct_play
+    options["play_random"] = args.play_random_station
 
     options["sort_by"] = args.stations_sort_by
+    options["filter_with"] = args.stations_filter_with
 
     options["discover_country_code"] = args.discover_country_code
     options["discover_state"] = args.discover_state
@@ -53,5 +55,6 @@ def parse_options():
 
     options["target_url"] = ""
     options["volume"] = args.volume
+    options["audio_player"] = args.audio_player
 
     return options